+++ x/name/keeper/msg_server.go
@@ -191,9 +191,7 @@ func (s msgServer) CreateRootName(goCtx context.Context, msg *types.MsgCreateRoo
 		return nil, errors.Wrapf(govtypes.ErrInvalidSigner, "expected %s got %s", s.Keeper.GetAuthority(), msg.Authority)
 	}
 
-	// Routes to legacy proposal handler to avoid code duplication
-	// Setting title and description to empty strings. These two fields are deprecated in the v1.
-	err := HandleCreateRootNameProposal(ctx, s.Keeper, types.NewCreateRootNameProposal("", "", msg.Record.Name, sdk.AccAddress(msg.Record.Address), msg.Record.Restricted))
+	err := s.Keeper.CreateRootName(ctx, msg.Record.Name, msg.Record.Address, msg.Record.Restricted)
 	if err != nil {
 		return nil, err
 	}